size_t -> unsigned int

+++ src/crypto/crypto-ops.c
@@ -335,7 +335,7 @@ static int fe_isnonzero(const fe f) {
 
 // Montgomery's trick
 // https://iacr.org/archive/pkc2004/29470042/29470042.pdf 2.2
-int fe_batch_invert(fe *out, const fe *in, const size_t n) {
+int fe_batch_invert(fe *out, const fe *in, const unsigned int n) {
   if (n == 0) {
     return 0;
   }
@@ -353,7 +353,7 @@ int fe_batch_invert(fe *out, const fe *in, const size_t n) {
 
   // Step 1: collect initial muls
   fe_copy(out[0], in[0]);
-  for (size_t i = 1; i < n; ++i) {
+  for (unsigned int i = 1; i < n; ++i) {
     fe_mul(out[i], out[i-1], in[i]);
   }
 
@@ -366,7 +366,7 @@ int fe_batch_invert(fe *out, const fe *in, const size_t n) {
   fe_invert(a, out[n-1]);
 
   // Step 3: get each inverse
-  for (size_t i = n; i > 1; --i) {
+  for (unsigned int i = n; i > 1; --i) {
     fe_mul(out[i-1], a, out[i-2]);
     fe_mul(a, a, in[i-1]);
   }
