+++ src/crypto/crypto-ops.h
@@ -30,7 +30,6 @@
 
 #pragma once
 
-#include <stddef.h>
 #include <stdint.h>
 
 /* From fe.h */
@@ -178,7 +177,7 @@ If `out` and `in` overlap, it will cause undefined output.
 
 No 0 fe's are expected for `in`, otherwise fails.
 **/
-int fe_batch_invert(fe *out, const fe *in, const size_t n);
+int fe_batch_invert(fe *out, const fe *in, const unsigned int n);
 void fe_mul(fe out, const fe, const fe);
 void fe_neg(fe h, const fe f);
 void fe_sq(fe h, const fe f);
